refactor: enhance code clarity and consistency across renderer files
- Updated parameter types in several functions to use const references for improved performance and clarity.
- Standardized the use of `const` for local variables and function parameters where applicable.
- Cleaned up includes and improved organization in various rendering components, ensuring consistent naming and formatting.
- Enhanced readability of the codebase by simplifying variable declarations and initializations.

+++ engine/renderer/fullscreen_quad.cpp
@@ -23,8 +23,7 @@ namespace RealmEngine
         if (!m_impl || !m_impl->mesh.vertex_input)
             return;
 
-        m_impl->mesh.vertex_input->draw(PrimitiveType::Triangles,
-                                        static_cast<uint32_t>(PrimitiveVertices::k_fullscreen_quad_vertex_count));
+        m_impl->mesh.vertex_input->draw(PrimitiveType::Triangles, PrimitiveVertices::k_fullscreen_quad_vertex_count);
     }
 
 } // namespace RealmEngine
